refactor(auth): move profile-related logic to profile module
- Removed `profile` and `update-profile` endpoints from `AuthController` and `AuthService`.
- Added corresponding `profile` and `update-profile` endpoints to `ProfileController` and `ProfileService`.
- Enhanced `ProfileService` with dependency injection for `UsersService`.
- Updated `ProfileModule` to include required imports and providers (`UsersService`, `PrismaService`, `TokenService`).
